+++ packages/react/components/chips/list/ChipsList.tsx
@@ -3,7 +3,7 @@ import { hashClass } from '@/helpers/hashClassesHelpers'
 import { is } from '@/services/classify'
 import clsx from 'clsx'
 import * as React from 'react'
-import { ChipsListProps } from './ChipsListProps'
+import { ChipsListProps, ChipsListRef } from './ChipsListProps'
 
 /**
  * ChipsList Component - Container for Chips
@@ -14,15 +14,17 @@ import { ChipsListProps } from './ChipsListProps'
  * @param scrollable {boolean} If multiple Chips make scrollable List
  * @param others
  */
-const ChipsList = ({ className, id, children, multiple, scrollable, ...others }: ChipsListProps) => {
-  const classes = hashClass(clsx('chips-list', multiple && is('multiple'), scrollable && is('scrollable'), className))
+const ChipsList = React.forwardRef<ChipsListRef, ChipsListProps>(
+  ({ className, id, children, multiple, scrollable, accessibilityLabelledBy, ...others }, ref) => {
+    const classes = hashClass(clsx('chips-list', multiple && is('multiple'), scrollable && is('scrollable'), className))
 
-  return (
-    <div id={id} role='group' className={classes} {...others}>
-      {children}
-    </div>
-  )
-}
+    return (
+      <div ref={ref} id={id} role='group' aria-labelledby={accessibilityLabelledBy} className={classes} {...others}>
+        {children}
+      </div>
+    )
+  },
+)
 
 ChipsList.displayName = ComponentName.ChipsList
 export default ChipsList