+++ packages/react/components/tabs/hooks/useTabs.tsx
@@ -4,9 +4,10 @@ import { TabsContext } from '../context'
 interface IParams {
   activeIndex?: number
   inverted?: boolean
+  small?: boolean
 }
 
-export const useTabs = ({ activeIndex, inverted }: IParams) => {
+export const useTabs = ({ activeIndex, inverted, small }: IParams) => {
   try {
     const [currentIndex, setCurrentIndex] = React.useState<number>(activeIndex || 0)
     const [isInverted, setIsInverted] = React.useState<boolean>(inverted || false)
@@ -19,6 +20,7 @@ export const useTabs = ({ activeIndex, inverted }: IParams) => {
       return (
         <TabsContext.Provider
           value={{
+            small,
             activeIndex: currentIndex,
             inverted: isInverted,
             setInverted: setIsInverted,